Refactor api calls into home and apply localStorage
